Removed miscellaneous common folder.
