--- common/steamcraft/client/lib/ClientResources.java
@@ -1,34 +1,0 @@
-/**
- * This class was created by <MrArcane111> or his SC2 development team. 
- * This class is available as part of the Steamcraft 2 Mod for Minecraft.
- *
- * Steamcraft 2 is open-source and is distributed under the MMPL v1.0 License.
- * (http://www.mod-buildcraft.com/MMPL-1.0.txt)
- * 
- * Steamcraft 2 is based on the original Steamcraft created by Proloe.
- * Steamcraft (c) Proloe 2011
- * (http://www.minecraftforum.net/topic/251532-181-steamcraft-source-code-releasedmlv054wip/)
- * 
- * Some code is derived from PowerCraft created by MightyPork which is registered
- * under the MMPL v1.0.
- * PowerCraft (c) MightyPork 2012
- *
- * File created @ [Feb 4, 2014, 5:19:04 PM]
- */
-package common.steamcraft.client.lib2;
-
-/**
- * @author MrArcane111 & general3214
- *
- */
-public class ClientResources { // TODO: Please move this to ClientHelper.java
-	public static final String PREFIX_MOD = "steamcraft:";
-
-    //public static final String PREFIX_LANG = "/assets/steamcraft/lang/";
-
-    public static final String PREFIX_GUI = PREFIX_MOD + "textures/gui/";
-    public static final String PREFIX_MISC = PREFIX_MOD + "textures/misc/";
-    public static final String PREFIX_MODEL = PREFIX_MOD + "textures/model/";
-
-    public static final String EMPTY_TEXTURE = "emptyTexture";
-}
