--- common/steamcraft/common/lib/BlockIDs.java
@@ -1,67 +1,0 @@
-package common.steamcraft.common.lib2;
-
-import common.steamcraft.common.SC2;
-
-public class BlockIDs {
-    //public static int torchElectricIdleID = 2600;
-    //public static int torchElectricActiveID = 2601;
-    //public static int torchTeslaIdleID = 2602;
-    //public static int torchTeslaActiveID = 2603;
-    //public static int teslaReceiverIdleID = 2604;
-    //public static int teslaReceiverActiveID = 2605;
-    //public static int batteryID = 2612;
-    public static int steamOvenIdleID = SC2.proxy.cfgHandler.loadBlock("steamOvenIdleID");
-    public static int steamOvenActiveID = SC2.proxy.cfgHandler.loadBlock("steamOvenActiveID");
-    public static int chemOvenIdleID = SC2.proxy.cfgHandler.loadBlock("chemOvenIdleID");
-    public static int chemOvenActiveID = SC2.proxy.cfgHandler.loadBlock("chemOvenActiveID");
-    public static int nukeOvenIdleID = SC2.proxy.cfgHandler.loadBlock("nukeOvenIdleID");
-    public static int nukeOvenActiveID = SC2.proxy.cfgHandler.loadBlock("nukeOvenActiveID");
-    public static int oreBrimstoneID = SC2.proxy.cfgHandler.loadBlock("oreBrimstoneID");
-    public static int oreBorniteID = SC2.proxy.cfgHandler.loadBlock("oreBorniteID");
-    public static int orePhosphateID = SC2.proxy.cfgHandler.loadBlock("orePhosphateID");
-    public static int oreUraniteID = SC2.proxy.cfgHandler.loadBlock("oreUraniteID");
-    public static int oreVoluciteID = SC2.proxy.cfgHandler.loadBlock("oreVoluciteID");
-    public static int oreAluminumID = SC2.proxy.cfgHandler.loadBlock("oreAluminumID");
-    public static int oreCopperID = SC2.proxy.cfgHandler.loadBlock("oreCopperID");
-    public static int oreTinID = SC2.proxy.cfgHandler.loadBlock("oreTinID");
-    public static int oreZincID = SC2.proxy.cfgHandler.loadBlock("oreZincID");
-    public static int blockCastIronID = SC2.proxy.cfgHandler.loadBlock("blockCastIronID");
-    public static int blockVoluciteID = SC2.proxy.cfgHandler.loadBlock("blockVoluciteID");
-    public static int blockBrassID = SC2.proxy.cfgHandler.loadBlock("blockBrassID");
-    public static int blockUraniumID = SC2.proxy.cfgHandler.loadBlock("blockUraniumID");
-    public static int blockAluminumID = SC2.proxy.cfgHandler.loadBlock("blockAluminumID");
-    public static int blockCopperID = SC2.proxy.cfgHandler.loadBlock("blockCopperID");
-    public static int blockTinID = SC2.proxy.cfgHandler.loadBlock("blockTinID");
-    public static int blockZincID = SC2.proxy.cfgHandler.loadBlock("blockZincID");
-    public static int blockBronzeID = SC2.proxy.cfgHandler.loadBlock("blockBronzeID");
-    public static int blockSteelID = SC2.proxy.cfgHandler.loadBlock("blockSteelID");
-    public static int decorIronID = SC2.proxy.cfgHandler.loadBlock("decorIronID");
-    public static int decorGoldID = SC2.proxy.cfgHandler.loadBlock("decorGoldID");
-    public static int decorDiamondID = SC2.proxy.cfgHandler.loadBlock("decorDiamondID");
-    public static int decorCastIronID = SC2.proxy.cfgHandler.loadBlock("decorCastIronID");
-    public static int decorBrassID = SC2.proxy.cfgHandler.loadBlock("decorBrassID");
-    public static int decorVoluciteID = SC2.proxy.cfgHandler.loadBlock("decorVoluciteID");
-    public static int decorLapisID = SC2.proxy.cfgHandler.loadBlock("decorLapisID");
-    public static int decorStoneID = SC2.proxy.cfgHandler.loadBlock("decorStoneID");
-    public static int decorUraniumID = SC2.proxy.cfgHandler.loadBlock("decorUraniumID");
-    public static int railingCastIronID = SC2.proxy.cfgHandler.loadBlock("railingCastIronID");
-    public static int gateCastIronID = SC2.proxy.cfgHandler.loadBlock("gateCastIronID");
-    public static int lampOnID = SC2.proxy.cfgHandler.loadBlock("lampOnID");
-    public static int lampOffID = SC2.proxy.cfgHandler.loadBlock("lampOffID");
-    public static int woodBrassID = SC2.proxy.cfgHandler.loadBlock("woodBrassID");
-    public static int leavesLampID = SC2.proxy.cfgHandler.loadBlock("leavesLampID");
-    public static int ironLampIdleID = SC2.proxy.cfgHandler.loadBlock("ironLampIdleID");
-    public static int ironLampActiveID = SC2.proxy.cfgHandler.loadBlock("ironLampActiveID");
-    public static int teaPlantID = SC2.proxy.cfgHandler.loadBlock("teaPlantID");
-    public static int castIronLampIdleID = SC2.proxy.cfgHandler.loadBlock("castIronLampIdleID");
-    public static int castIronLampActiveID = SC2.proxy.cfgHandler.loadBlock("castIronLampActiveID");
-    public static int copperWireID = SC2.proxy.cfgHandler.loadBlock("copperWireID");
-    public static int blockSlateTileID = SC2.proxy.cfgHandler.loadBlock("blockSlateTileID");
-    public static int stairsSlateTileID = SC2.proxy.cfgHandler.loadBlock("stairsSlateTileID");
-    public static int lightningRodID = SC2.proxy.cfgHandler.loadBlock("lightningRodID");
-	public static int compressorID = SC2.proxy.cfgHandler.loadBlock("compressorID");
-	public static int coveyorID = SC2.proxy.cfgHandler.loadBlock("conveyorID");
-	public static int generatorID = SC2.proxy.cfgHandler.loadBlock("generatorID");
-	public static int smogID = SC2.proxy.cfgHandler.loadBlock("smogID");
-	public static int etheriumCrystalID = SC2.proxy.cfgHandler.loadBlock("etheriumCrystalID");
-}
