--- common/steamcraft/common/lib/BlockNames.java
@@ -1,63 +1,0 @@
-/**
- * This class was created by <MrArcane111> or his SC2 development team. 
- * This class is available as part of the Steamcraft 2 Mod for Minecraft.
- *
- * Steamcraft 2 is open-source and is distributed under the MMPL v1.0 License.
- * (http://www.mod-buildcraft.com/MMPL-1.0.txt)
- * 
- * Steamcraft 2 is based on the original Steamcraft created by Proloe.
- * Steamcraft (c) Proloe 2011
- * (http://www.minecraftforum.net/topic/251532-181-steamcraft-source-code-releasedmlv054wip/)
- * 
- * File created @ [Jan 30, 2014, 6:11:43 PM]
- */
-package common.steamcraft.common.lib2;
-
-/**
- * @author MrArcane111
- *
- */
-public class BlockNames 
-{
-	public static final String ORE_BORNITE = LibInfo.SC2_PREFIX + "bornite";
-	public static final String ORE_BRIMSTONE = LibInfo.SC2_PREFIX + "brimstone";
-	public static final String ORE_PHOSPHATE = LibInfo.SC2_PREFIX + "phosphate";
-	public static final String ORE_URANITE = LibInfo.SC2_PREFIX + "uranite";
-	public static final String ORE_VOLUCITE = LibInfo.SC2_PREFIX + "volucite";
-	public static final String ORE_ALUMINUM = LibInfo.SC2_PREFIX + "aluminum";
-	public static final String ORE_COPPER = LibInfo.SC2_PREFIX + "copper";
-	public static final String ORE_TIN = LibInfo.SC2_PREFIX + "tin";
-	public static final String ORE_ZINC = LibInfo.SC2_PREFIX + "zinc";
-
-	public static final String COPPER_WIRE = LibInfo.SC2_PREFIX + "redstonedust";
-	public static final String LIGHTNING_ROD = LibInfo.SC2_PREFIX + "lightningrod";
-
-	public static final String BLOCK_CAST_IRON = LibInfo.SC2_PREFIX + "castironblock";
-	public static final String BLOCK_VOLICITE = LibInfo.SC2_PREFIX + "volucite";
-	public static final String BLOCK_BRASS = LibInfo.SC2_PREFIX + "brassblock";
-	public static final String BLOCK_URANIUM = LibInfo.SC2_PREFIX + "uraniumblock";
-	public static final String BLOCK_ALUMINUM = LibInfo.SC2_PREFIX + "aluminumblock";
-	public static final String BLOCK_COPPER = LibInfo.SC2_PREFIX + "copperblock";
-	public static final String BLOCK_TIN = LibInfo.SC2_PREFIX + "tinblock";
-	public static final String BLOCK_ZINC = LibInfo.SC2_PREFIX + "zincblock";
-	public static final String BLOCK_SLATE_TILES = LibInfo.SC2_PREFIX + "black_slatetiles"; // Looks like satellites
-	public static final String STAIRS_SLATE_TILES = LibInfo.SC2_PREFIX + "black_slatetiles";
-	public static final String DECOR_IRON = LibInfo.SC2_PREFIX + "engriron";
-	public static final String DECOR_GOLD = LibInfo.SC2_PREFIX + "engrgold";
-	public static final String DECOR_DIAMOND = LibInfo.SC2_PREFIX + "engrdiamond";
-	public static final String DECOR_CAST_IRON = LibInfo.SC2_PREFIX + "engrcastiron";
-	public static final String DECOR_BRASS = LibInfo.SC2_PREFIX + "engrbrass";
-	public static final String DECOR_VOLUCITE = LibInfo.SC2_PREFIX + "engrvolucite";
-	public static final String DECOR_LAPIS = LibInfo.SC2_PREFIX + "engrlapis";
-	public static final String DECOR_STONE = LibInfo.SC2_PREFIX + "carvedstone";
-	public static final String DECOR_URANIUM = LibInfo.SC2_PREFIX + "engruranium";
-	public static final String RAILING_CAST_IRON = LibInfo.SC2_PREFIX + "castironblock";
-	public static final String GATE_CAST_IRON = LibInfo.SC2_PREFIX + "castironblock";
-	public static final String LAMP_ON = LibInfo.SC2_PREFIX + "lampblockon";
-	public static final String LAMP_OFF = LibInfo.SC2_PREFIX + "lampblockoff";
-	public static final String WOOD_BRASS = LibInfo.SC2_PREFIX + "brasswood";
-	public static final String LEAVES_BRASS = LibInfo.SC2_PREFIX + "brassleaves";
-	public static final String PLANT_TEA = LibInfo.SC2_PREFIX + "tea";
-	public static final String CAST_IRON_LAMP_IDLE = LibInfo.SC2_PREFIX + "ironlamp";
-	public static final String CAST_IRON_LAMP_ACTIVE = LibInfo.SC2_PREFIX + "ironlamp";
-}
