--- common/steamcraft/common/lib/EntityIDs.java
@@ -1,9 +1,0 @@
-package common.steamcraft.common.lib2;
-
-import common.steamcraft.common.SC2;
-
-public class EntityIDs
-{
-	public static final int BULLET_ID = SC2.proxy.cfgHandler.loadEntity("BULLET_ID");
-	public static final int LIGHTNINGROD_ID = SC2.proxy.cfgHandler.loadEntity("LIGHTNINGROD_ID");
-}
