--- common/steamcraft/common/lib/ItemIDs.java
@@ -1,95 +1,0 @@
-package common.steamcraft.common.lib2;
-
-import common.steamcraft.common.SC2;
-
-public class ItemIDs
-{
-    public static int etheriumID = SC2.proxy.cfgHandler.loadItem("etheriumID");
-    public static int chemicSaltID = SC2.proxy.cfgHandler.loadItem("chemicSaltID");
-    public static int borniteID = SC2.proxy.cfgHandler.loadItem("borniteID");
-    public static int obsidianSlateID = SC2.proxy.cfgHandler.loadItem("obsidianSlateID");
-    public static int ingotBrassID = SC2.proxy.cfgHandler.loadItem("ingotBrassID");
-    public static int ingotCastIronID = SC2.proxy.cfgHandler.loadItem("ingotCastIronID");
-	public static int ingotAluminumID = SC2.proxy.cfgHandler.loadItem("ingotAluminumID");
-	public static int ingotCopperID = SC2.proxy.cfgHandler.loadItem("ingotCopperID");
-	public static int ingotTinID = SC2.proxy.cfgHandler.loadItem("ingotTinID");
-	public static int ingotZincID = SC2.proxy.cfgHandler.loadItem("ingotZincID");
-	public static int ingotBronzeID = SC2.proxy.cfgHandler.loadItem("ingotBronzeID");
-	public static int ingotSteelID = SC2.proxy.cfgHandler.loadItem("ingotSteelID");
-    public static int lightBulbID = SC2.proxy.cfgHandler.loadItem("lightBulbID");
-    public static int phosphorusID = SC2.proxy.cfgHandler.loadItem("phosphorusID");
-    public static int uraniumStoneID = SC2.proxy.cfgHandler.loadItem("uraniumStoneID");
-    public static int uraniumID = SC2.proxy.cfgHandler.loadItem("uraniumID");
-    public static int reactorCoreID = SC2.proxy.cfgHandler.loadItem("reactorCoreID");
-    public static int coreDrillID = SC2.proxy.cfgHandler.loadItem("coreDrillID");
-    public static int drillBaseID = SC2.proxy.cfgHandler.loadItem("drillBaseID");
-    public static int pickaxeObsidianID = SC2.proxy.cfgHandler.loadItem("pickaxeObsidianID");
-    public static int shovelObsidianID = SC2.proxy.cfgHandler.loadItem("shovelObsidianID");
-    public static int axeObsidianID = SC2.proxy.cfgHandler.loadItem("axeObsidianID");
-    public static int hoeObsidianID = SC2.proxy.cfgHandler.loadItem("hoeObsidianID");
-    public static int swordObsidianID = SC2.proxy.cfgHandler.loadItem("swordObsidianID");
-    public static int drillObsidianID = SC2.proxy.cfgHandler.loadItem("drillObsidianID");
-    public static int helmetObsidianID = SC2.proxy.cfgHandler.loadItem("helmetObsidianID");
-    public static int plateObsidianID = SC2.proxy.cfgHandler.loadItem("plateObsidianID");
-    public static int legsObsidianID = SC2.proxy.cfgHandler.loadItem("legsObsidianID");
-    public static int bootsObsidianID = SC2.proxy.cfgHandler.loadItem("bootsObsidianID");
-    public static int pickaxeEtheriumID = SC2.proxy.cfgHandler.loadItem("pickaxeEtheriumID");
-    public static int shovelEtheriumID = SC2.proxy.cfgHandler.loadItem("shovelEtheriumID");
-    public static int axeEtheriumID = SC2.proxy.cfgHandler.loadItem("axeEtheriumID");
-    public static int hoeEtheriumID = SC2.proxy.cfgHandler.loadItem("hoeEtheriumID");
-    public static int swordEtheriumID = SC2.proxy.cfgHandler.loadItem("swordEtheriumID");
-    public static int drillEtheriumID = SC2.proxy.cfgHandler.loadItem("drillEtheriumID");
-    public static int pickaxeSteamID = SC2.proxy.cfgHandler.loadItem("pickaxeSteamID");
-    public static int shovelSteamID = SC2.proxy.cfgHandler.loadItem("shovelSteamID");
-    public static int axeSteamID = SC2.proxy.cfgHandler.loadItem("axeSteamID");
-    public static int hoeSteamID = SC2.proxy.cfgHandler.loadItem("hoeSteamID");
-    public static int swordSteamID = SC2.proxy.cfgHandler.loadItem("swordSteamID");
-    public static int drillSteamID = SC2.proxy.cfgHandler.loadItem("drillSteamID");
-    public static int drillWoodID = SC2.proxy.cfgHandler.loadItem("drillWoodID");
-    public static int drillStoneID = SC2.proxy.cfgHandler.loadItem("drillStoneID");
-    public static int drillIronID = SC2.proxy.cfgHandler.loadItem("drillIronID");
-    public static int drillGoldID = SC2.proxy.cfgHandler.loadItem("drillGoldID");
-    public static int drillDiamondID = SC2.proxy.cfgHandler.loadItem("drillDiamondID");
-    public static int chiselID = SC2.proxy.cfgHandler.loadItem("chiselID");
-    public static int spannerID = SC2.proxy.cfgHandler.loadItem("spannerID");
-    public static int musketRoundID = SC2.proxy.cfgHandler.loadItem("musketRoundID");
-    public static int percussionRoundID = SC2.proxy.cfgHandler.loadItem("percussionRoundID");
-    public static int percussionLockID = SC2.proxy.cfgHandler.loadItem("percussionLockID");
-    public static int smoothBarrelID = SC2.proxy.cfgHandler.loadItem("smoothBarrelID");
-    public static int rifledBarrelID = SC2.proxy.cfgHandler.loadItem("rifledBarrelID");
-    public static int woodenStockID = SC2.proxy.cfgHandler.loadItem("woodenStockID");
-    public static int flintlockMusketID = SC2.proxy.cfgHandler.loadItem("flintlockMusketID");
-    public static int matchlockMusketID = SC2.proxy.cfgHandler.loadItem("matchlockMusketID");
-    public static int percussionCapMusketID = SC2.proxy.cfgHandler.loadItem("percussionCapMusketID");
-    public static int flintlockRifleID = SC2.proxy.cfgHandler.loadItem("flintlockRifleID");
-    public static int matchlockRifleID = SC2.proxy.cfgHandler.loadItem("matchlockRifleID");
-    public static int percussionCapRifleID = SC2.proxy.cfgHandler.loadItem("percussionCapRifleID");
-    public static int ironLampID = SC2.proxy.cfgHandler.loadItem("ironLampID");
-    public static int teaSeedID = SC2.proxy.cfgHandler.loadItem("teaSeedID");
-    public static int teaLeavesID = SC2.proxy.cfgHandler.loadItem("teaLeavesID");
-    public static int coldKettleID = SC2.proxy.cfgHandler.loadItem("coldKettleID");
-    public static int hotKettleID = SC2.proxy.cfgHandler.loadItem("hotKettleID");
-    public static int emptyKettleID = SC2.proxy.cfgHandler.loadItem("emptyKettleID");
-    public static int fullTeacupID = SC2.proxy.cfgHandler.loadItem("fullTeacupID");
-    public static int emptyTeacupID = SC2.proxy.cfgHandler.loadItem("emptyTeacupID");
-    public static int brassGogglesID = SC2.proxy.cfgHandler.loadItem("brassGogglesID");
-    public static int aqualungID = SC2.proxy.cfgHandler.loadItem("aqualungID");
-    public static int legBracesID = SC2.proxy.cfgHandler.loadItem("legBracesID");
-    public static int rollerSkatesID = SC2.proxy.cfgHandler.loadItem("rollerSkatesID");
-    public static int helmetEtheriumID = SC2.proxy.cfgHandler.loadItem("helmetEtheriumID");
-    public static int plateEtheriumID = SC2.proxy.cfgHandler.loadItem("plateEtheriumID");
-    public static int legsEtheriumID = SC2.proxy.cfgHandler.loadItem("legsEtheriumID");
-    public static int bootsEtheriumID = SC2.proxy.cfgHandler.loadItem("bootsEtheriumID");
-    public static int rifleRoundID = SC2.proxy.cfgHandler.loadItem("rifleRoundID");
-    public static int copperWireID = SC2.proxy.cfgHandler.loadItem("copperWireID");
-    public static int rayGunID = SC2.proxy.cfgHandler.loadItem("rayGunID");
-    public static int brassWatchID = SC2.proxy.cfgHandler.loadItem("brassWatchID");
-    public static int brassWingsID = SC2.proxy.cfgHandler.loadItem("brassWingsID");
-    public static int brassMonocleID = SC2.proxy.cfgHandler.loadItem("brassMonocleID");
-	public static int steamCanisterID = SC2.proxy.cfgHandler.loadItem("steamCanisterID");
-	public static int gasCanisterID = SC2.proxy.cfgHandler.loadItem("gasCanisterID");
-	public static int capeID = SC2.proxy.cfgHandler.loadItem("capeID");
-	public static int utilityBeltID = SC2.proxy.cfgHandler.loadItem("utilityBeltID");
-	public static int emptyCanisterID = SC2.proxy.cfgHandler.loadItem("emptyCanisterID");
-    public static int guideBookID = SC2.proxy.cfgHandler.loadItem("guideBookID");
-}
