--- common/steamcraft/common/lib/ItemNames.java
@@ -1,10 +1,0 @@
-package common.steamcraft.common.lib2;
-
-public class ItemNames 
-{
-	public static final String ETHERIUM = LibInfo.SC2_PREFIX + "etherium";
-
-	public static final String COMPRESSED_GAS = LibInfo.SC2_PREFIX + "compressedgas";
-
-	public static final String LEATHER_CAPE = LibInfo.SC2_PREFIX + "cape";
-}
