+++ src/main/java/steamcraft/common/lib/EntityIDs.java
@@ -1,0 +1,8 @@
+package steamcraft.common.lib;
+
+
+public class EntityIDs
+{
+	//public static final int BULLET_ID = SC2.proxy.cfgHandler.loadEntity("BULLET_ID");
+	//public static final int LIGHTNINGROD_ID = SC2.proxy.cfgHandler.loadEntity("LIGHTNINGROD_ID");
+}
